requirements.txt
flask
flask-cors
newspaper3k
google-cloud-texttospeech